refactor: keep reward bucket types local
